MOVEONLY: Header verification to consensus/header_verify.o
