+++ src/Makefile.am
@@ -95,6 +95,7 @@ BITCOIN_CORE_H = \
   compat/sanity.h \
   compressor.h \
   consensus/consensus.h \
+  consensus/header_verify.h \
   consensus/tx_verify.h \
   core_io.h \
   core_memusage.h \
@@ -186,6 +187,7 @@ libbitcoin_server_a_SOURCES = \
   blockencodings.cpp \
   chain.cpp \
   checkpoints.cpp \
+  consensus/header_verify.cpp \
   consensus/tx_verify.cpp \
   httprpc.cpp \
   httpserver.cpp \
